chore: merge LoadFrom components
